backend/requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==10.4.0
environs[django]==9.3.2
phonenumbers==8.13.42
django-phonenumber-field==8.0.0
djangorestframework==3.13.1
requests==2.32.3
geopy==2.4.1
rollbar==1.0.0
psycopg2==2.9.10
gunicorn==23.0.0
